fixed pages that was jsx files so that they are tsx, fixed type issues and some semantics

+++ src/pages/profile/index.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from "react";
+import React, { useState, useEffect, FormEvent } from "react";
 import LoginForm from "../../components/forms/loginForm";
 import RegisterForm from "../../components/forms/registerForm";
 import { useAuth } from "../../context/authContext";
@@ -17,27 +17,29 @@ import {
 } from "../../helpers/handlers";
 import BookingLi from "../../components/profileContent/bookingLi";
 import { FaPlus } from "react-icons/fa6";
+import { Venue, Booking } from "../../types";
 
 function Profile() {
   const { user } = useAuth();
-  const [showRegisterForm, setShowRegisterForm] = useState(false);
+  const [showRegisterForm, setShowRegisterForm] = useState<boolean>(false);
   const toggleForm = () => setShowRegisterForm((prev) => !prev);
-  const [showModal, setShowModal] = useState(false);
-  const [selectedVenueId, setSelectedVenueId] = useState();
-  const [venues, setVenues] = useState([]);
-  const [userBookings, setUserBookings] = useState([]);
+  const [showModal, setShowModal] = useState<boolean>(false);
+  const [selectedVenueId, setSelectedVenueId] = useState<string | null>(null);
+  const [venues, setVenues] = useState<Venue[]>([]);
+  const [userBookings, setUserBookings] = useState<Booking[]>([]);
   const [formData, handleChange, resetFormData] = useVenueForm(selectedVenueId);
-  const [feedbackMessage, setFeedbackMessage] = useState("");
-  const [showFeedbackMessage, setShowFeedbackMessage] = useState(false);
-  const [isSuccessMessage, setIsSuccessMessage] = useState(false);
-  const [updatedVenueId, setUpdatedVenueId] = useState(null);
-  const [showConfirmModal, setShowConfirmModal] = useState(false);
-  const [venueToDelete, setVenueToDelete] = useState(null);
+  const [feedbackMessage, setFeedbackMessage] = useState<string>("");
+  const [showFeedbackMessage, setShowFeedbackMessage] =
+    useState<boolean>(false);
+  const [isSuccessMessage, setIsSuccessMessage] = useState<boolean>(false);
+  const [updatedVenueId, setUpdatedVenueId] = useState<string | null>(null);
+  const [showConfirmModal, setShowConfirmModal] = useState<boolean>(false);
+  const [venueToDelete, setVenueToDelete] = useState<string | null>(null);
   const navigate = useNavigate();
 
   const toggleModal = () => setShowModal((prev) => !prev);
 
-  const openModal = (venueId) => {
+  const openModal = (venueId: string | null = null) => {
     setSelectedVenueId(venueId);
     if (!venueId) {
       resetFormData();
@@ -50,7 +52,7 @@ function Profile() {
     setSelectedVenueId(null);
   };
 
-  const handleSubmit = async (event) => {
+  const handleSubmit = async (event: FormEvent) => {
     event.preventDefault();
     if (selectedVenueId) {
       await handleUpdate(event, selectedVenueId, formData);
@@ -61,43 +63,55 @@ function Profile() {
       setTimeout(() => setShowFeedbackMessage(false), 1000);
       closeModal();
     } else {
-      await handleCreate(event, formData, (newVenue) => {
+      await handleCreate(event, formData, (newVenue: Venue) => {
         setVenues((prevVenues) => [newVenue, ...prevVenues]);
         setFeedbackMessage("New");
         setIsSuccessMessage(true);
         setShowFeedbackMessage(true);
-        setUpdatedVenueId(newVenue.id);
+        if (newVenue.id) {
+          setUpdatedVenueId(newVenue.id);
+        }
         setTimeout(() => setShowFeedbackMessage(false), 1000);
         closeModal();
       });
     }
   };
 
-  const handleClick = (id) => {
-    console.log(`Navigating to venue with ID: ${id}`);
-    navigate(`/venue/${id}`);
+  const handleClick = (id: string | undefined) => {
+    if (id) {
+      console.log(`Navigating to venue with ID: ${id}`);
+      navigate(`/venue/${id}`);
+    }
   };
 
-  const handleDeleteClick = (venueId) => {
-    setVenueToDelete(venueId);
-    setShowConfirmModal(true);
+  const handleDeleteClick = (venueId: string | undefined) => {
+    if (venueId) {
+      setVenueToDelete(venueId);
+      setShowConfirmModal(true);
+    }
   };
 
   const confirmDelete = async () => {
-    await handleDelete(venueToDelete, setVenues, venues);
-    setShowConfirmModal(false);
+    if (venueToDelete) {
+      await handleDelete(venueToDelete, setVenues, venues);
+      setShowConfirmModal(false);
+    }
   };
 
   const fetchBookings = async () => {
     try {
-      const bookingResponse = await profileAPI.fetchProfileBookings(
-        user.name,
-        "true"
-      );
-      const sortedBookings = bookingResponse.data.data.bookings.sort(
-        (a, b) => new Date(b.created) - new Date(a.created)
-      );
-      setUserBookings(sortedBookings);
+      if (user?.name) {
+        const bookingResponse = await profileAPI.fetchProfileBookings(
+          user.name,
+          "true"
+        );
+        const bookings = bookingResponse.data.data.bookings || [];
+        const sortedBookings = bookings.sort(
+          (a: Booking, b: Booking) =>
+            new Date(b.dateFrom).getTime() - new Date(a.dateFrom).getTime()
+        );
+        setUserBookings(sortedBookings);
+      }
     } catch (error) {
       console.error("Failed to fetch bookings:", error);
     }
@@ -108,7 +122,9 @@ function Profile() {
       if (user && user.name) {
         try {
           const response = await profileAPI.fetchProfile(user.name);
-          setVenues(response.data.data);
+          if (Array.isArray(response.data.data)) {
+            setVenues(response.data.data);
+          }
           await fetchBookings();
         } catch (error) {
           console.error("Failed to fetch data:", error);
@@ -131,8 +147,8 @@ function Profile() {
                   My Venues
                 </h2>
                 <PrimaryButton
-                  onClick={() => openModal(null)}
-                  className=" flex items-center gap-4 rounded-full shadow-lg py-2.5 px-5 mb-10 mt-5 bg-skin-infoBg text-skin-primary hover:bg-skin-primary hover:text-white text-lg"
+                  onClick={() => openModal()}
+                  className="flex items-center gap-4 rounded-full shadow-lg py-2.5 px-5 mb-10 mt-5 bg-skin-infoBg text-skin-primary hover:bg-skin-primary hover:text-white text-lg"
                   disabled={false}
                 >
                   <p>Add venue</p>
@@ -153,7 +169,7 @@ function Profile() {
                     key={venue.id}
                     venue={venue}
                     onClick={() => handleClick(venue.id)}
-                    onEdit={() => openModal(venue.id)}
+                    onEdit={() => openModal(venue.id || null)}
                     onDelete={() => handleDeleteClick(venue.id)}
                     feedbackMessage={feedbackMessage}
                     showFeedbackMessage={
@@ -201,12 +217,14 @@ function Profile() {
               <PrimaryButton
                 onClick={confirmDelete}
                 className="bg-red-500 hover:bg-red-600 px-2 sm:px-7 sm:py-2 text-white"
+                disabled={false}
               >
                 Delete
               </PrimaryButton>
               <PrimaryButton
                 onClick={() => setShowConfirmModal(false)}
-                className=" px-2 py-1 sm:px-7 sm:py-2 bg-gray-100 hover:bg-gray-200"
+                className="px-2 py-1 sm:px-7 sm:py-2 bg-gray-100 hover:bg-gray-200"
+                disabled={false}
               >
                 Cancel
               </PrimaryButton>
